fix: Notes not moving beyond hitLineY
The problem was caused by clamping of the progress to 1 if
it has passed the hitLineY, meaning the y of the Note is
now stuck to the hitLineY and doesn't move any further.

This fixes the bug by removing the maxY clamping so that the
y continues to be calculated until it gets removed once it
moves beyond the playfield.