feat: Implement calculateY() with no parameters in Note class
This is the start of the implementation of a new approach in calculating note position.
Instead of getting the parameters beforehand, the Note gets
updated, thus changing its y position.

It's cleaner this way because we can just use another class to
update the note (NoteManager) and another to render the note
 (PlayfieldRenderer).