feat: Add method to reset notes state
This is so that if we want to retry, the notes actually appear.
If the state isn't reset, all of the notes are marked either as
hit or miss. Meaning they don't appear in the playfield at all.

I hate my life, this is such a massive oversight that I spent
4 hours or more working on.